Week 7 - JPA를 이용한 db 단 레이어 구축

+++ community/src/main/java/kakaotech/community/domain/post/service/PostService.java
@@ -3,13 +3,17 @@ package kakaotech.community.domain.post.service;
 import kakaotech.community.domain.image.service.ImageService;
 import kakaotech.community.domain.post.Post;
 import kakaotech.community.domain.post.PostRepository;
+import kakaotech.community.domain.post.PostSummaryProjection;
 import kakaotech.community.domain.post.dto.PostResponse;
 import kakaotech.community.domain.postlike.service.PostLikeQueryService;
 import kakaotech.community.domain.user.User;
 import kakaotech.community.domain.user.service.UserService;
 import kakaotech.community.global.exception.PostException;
+import kakaotech.community.global.page.PageQuery;
+import kakaotech.community.global.page.PageResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -19,6 +23,7 @@ import static kakaotech.community.global.exception.code.ExceptionCode.POST_NOT_F
 import static kakaotech.community.global.exception.code.ExceptionCode.POST_WRITER_MISMATCH;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class PostService {
     private final UserService userService;
@@ -29,60 +34,72 @@ public class PostService {
 
     public PostResponse.Detail create(Long userId, String title, String content, MultipartFile image) {
         User user = userService.findById(userId);
+
+        if (image == null) {
+            Post post = postRepository.save(new Post(user, title, content));
+            return toDetail(user, post);
+        }
         UUID imageId = imageService.save(image);
 
-        Post post = postRepository.save(new Post(userId, title, content, imageId));
-
+        Post post = postRepository.save(new Post(user, title, content, imageId));
         return toDetail(user, post);
     }
 
-    public PostResponse.Detail getPost(Long postId) {
+    public PostResponse.Detail getPostDetail(Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new PostException(POST_NOT_FOUND));
 
         User writer = userService.findById(post.getWriterId());
 
         post.viewedOne();
-        postRepository.save(post);
 
         return toDetail(writer, post);
     }
 
-    public PostResponse.Summaries getPostsByPaging(int page) {
-        List<Post> posts = postRepository.findPostsByPaging(page);
+    @Transactional(readOnly = true)
+    public Post findById(Long postId) {
+        return postRepository.findById(postId)
+                .orElseThrow(() -> new PostException(POST_NOT_FOUND));
+    }
 
-        int totalSize = postRepository.size();
+    @Transactional(readOnly = true)
+    public PageResult<PostResponse.Summary> getPostsByPaging(int page) {
+        PageResult<PostSummaryProjection> postSummaries = postRepository.findPostsByPaging(PageQuery.offset(page));
 
-        return new PostResponse.Summaries(
-                posts.stream()
-                        .map(post -> {
-                                    User user = userService.findById(post.getWriterId());
+        return new PageResult<>(toSummariesResponse(postSummaries.elements()),
+                postSummaries.pageNum(), postSummaries.pageSize(), postSummaries.totalPage(), postSummaries.totalSize());
+    }
 
-                                    return PostMapper.toSummary(post, user);
-                                }
-                        ).toList(),
-                // totalSize 는 20개마다 1페이지씩 늘어남. (20개 -> 총 페이지 1, 39개 -> 1, 40개 -> 2 ...)
-                new PostResponse.Paging(page, 20, (totalSize % 20 == 0) ? totalSize / 20 : (totalSize / 20) + 1, totalSize)
-        );
+    private List<PostResponse.Summary> toSummariesResponse(List<PostSummaryProjection> projections) {
+        return projections.stream()
+                .map(p -> new PostResponse.Summary(
+                        p.postId(), p.title(), p.writerId(),
+                        p.writerName(), p.writerProfileImage(), p.likeCount(),
+                        p.commentCount(), p.viewCount(), p.createdAt()
+                )).toList();
     }
 
     private PostResponse.Detail toDetail(User user, Post post) {
-        boolean liked = postLikeQueryService.isLiked(post.getId(), user.getId());
+        boolean liked = postLikeQueryService.isLiked(post, user);
 
         return PostMapper.toDetail(user, post, liked);
     }
 
     public PostResponse.Detail update(Long userId, Long postId, String title, String content, MultipartFile image) {
-        Post prePost = postRepository.findById(postId)
-                .orElseThrow(() -> new PostException(POST_NOT_FOUND));
+        Post prePost = findById(postId);
 
         if (!prePost.isWrittenBy(userId)) {
             throw new PostException(POST_WRITER_MISMATCH);
         }
 
-        UUID newImageId = imageService.updateImage(prePost.getImageId(), image);
+        Post newPost;
+        if (image != null) {
+            UUID newImageId = imageService.updateImage(prePost.getImageId(), image);
+            newPost = prePost.update(title, content, newImageId);
+        } else {
+            newPost = prePost.update(title, content, null);
+        }
 
-        Post newPost = postRepository.save(prePost.update(title, content, newImageId));
         User user = userService.findById(userId);
 
         return toDetail(user, newPost);
@@ -96,10 +113,13 @@ public class PostService {
             throw new PostException(POST_WRITER_MISMATCH);
         }
 
-        imageService.delete(post.getImageId());
+        if (post.getImageId() != null) {
+            imageService.delete(post.getImageId());
+        }
         postRepository.deleteById(postId);
     }
 
+    @Transactional(readOnly = true)
     public void validatePost(Long postId) {
         if (!postRepository.existsById(postId)) {
             throw new PostException(POST_NOT_FOUND);
